+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chatbotkit/react
 
+## 1.22.0
+
+### Patch Changes
+
+- Updated dependencies [b17c1eb]
+  - @chatbotkit/sdk@1.22.0
+
 ## 1.21.10
 
 ### Patch Changes